better cache management

+++ docker-arch-with-req.bash
@@ -1,16 +1,18 @@
 #!/bin/bash
 
-set -eu -o pipefail
+set -eo pipefail
 
 PKGNAME="$(basename "$1")"
 
 rm -rf "$(pwd)/$PKGNAME-build"
 mkdir "$(pwd)/$PKGNAME-build"
 
+if [ "$CLEAN" == "1" ]; then
 pushd "$PKGNAME"
 git pull
 git clean -dfx
 popd
+fi
 
 pushd build-scripts
 
@@ -26,13 +28,15 @@ docker run \
   arch-amd64-builds \
   /build-scripts/build-with-req.bash "${@:2}"
 
-cp ../"$PKGNAME"/*.pkg.tar.* ../artifacts
+mv ../"$PKGNAME"/*.pkg.tar.* ../artifacts
 
 popd
 
 rm -rf "$(pwd)/$PKGNAME-build"
 
+if [ "$CLEAN" == "1" ]; then
 pushd "$PKGNAME"
 git pull
 git clean -dfx
 popd
+fi
